requirements.txt
pandas==0.23.4
numpy==1.16.2
matplotlib==2.2.3
praw==6.1.1
PyGithub==1.43.5
